refactor(nms): start shared/sequelize_models typescript migration
