+++ nms/shared/sequelize_models/models/audit_log_entry.js
@@ -15,8 +15,8 @@
  */
 
 import Sequelize from 'sequelize';
-
-import type {AssociateProp} from './AssociateTypes.flow';
+// $FlowFixMe migrated to typescript
+import type {AssociateProp} from './AssociateTypes';
 import type {DataTypes, Model} from 'sequelize';
 
 export type AuditLogEntryRawType = {
